move gfx tests to their own script
